routes: migrate preferences-config domain (10 routes)
Move to src/server/routes/preferences-config.ts:
- /api/config/cwd (GET, PUT)
- /api/preferences (GET, PUT)
- /api/project-config (GET, PUT)
- /api/project-config/defaults (GET)
- /api/config-directories (GET, DELETE, /reset POST)

handleApiRoute's PreferencesStore param renamed to _preferencesStore
(no remaining inner usage; outer createGateway scope still uses it).

Net: server.ts -183 lines (7236 → 7053). No behaviour change.
