+++ src/server/routes/dispatcher.ts
@@ -20,6 +20,7 @@ import { sandboxRoutes } from "./sandbox.js";
 import { oauthRoutes } from "./oauth.js";
 import { imageGenerationRoutes } from "./image-generation.js";
 import { modelsRoutes } from "./models.js";
+import { preferencesConfigRoutes } from "./preferences-config.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -31,6 +32,7 @@ const allRoutes: ReadonlyArray<Route> = [
 	...oauthRoutes,
 	...imageGenerationRoutes,
 	...modelsRoutes,
+	...preferencesConfigRoutes,
 ];
 
 /**
